requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
huggingface-hub>=0.20.3
numpy>=1.26.3
pydantic>=2.5.3
httpx>=0.26.0
python-multipart>=0.0.6

# Document parsing libraries
pandas>=2.0.0
openpyxl>=3.1.0
chardet>=5.0.0

# Note: Heavy ML libs (transformers, scikit-learn) removed for free tier
# Using Hugging Face Inference API instead of local models

